scrapper/requirements.txt
scrapy
scrapy-playwright
psycopg2-binary
